requirements.txt
# ── Broker integrations ────────────────────────────────────────────────────────
dhanhq>=2.0.2
fyers-apiv3>=3.1.11
robin_stocks==3.0.6
pyotp>=2.9.0
python-dotenv>=1.0.1

# ── Data & analysis ────────────────────────────────────────────────────────────
yfinance>=0.2.40
pandas>=2.2.2
numpy>=1.26.4
scipy>=1.13.0
statsmodels>=0.14.2
pandas_datareader==0.10.0
numpy_financial>=1.0.0
python_dateutil>=2.9.0
ta>=0.11.0

# ── Machine learning ───────────────────────────────────────────────────────────
# tensorflow 2.16+ required for Python 3.12
tensorflow>=2.16.1
# fbprophet is EOL — replaced by prophet
prophet>=1.1.5
scikit_learn>=1.4.2
pmdarima>=2.0.4
factor_analyzer>=0.5.1

# ── Deep learning / AI ─────────────────────────────────────────────────────────
keras>=3.3.3
fastai>=2.7.15

# ── Visualisation ──────────────────────────────────────────────────────────────
matplotlib>=3.8.4
seaborn>=0.13.2
mplfinance==0.12.10b0
networkx>=3.3
wordcloud>=1.9.3

# ── Web / scraping ─────────────────────────────────────────────────────────────
Requests>=2.31.0
beautifulsoup4>=4.12.3
lxml>=5.2.2
selenium>=4.21.0
webdriver_manager>=4.0.1
autoscraper==1.1.14
Flask>=3.0.3

# ── APIs & social data ─────────────────────────────────────────────────────────
praw>=7.7.1
tweepy>=4.14.0
twilio>=9.2.3
FundamentalAnalysis==0.3.1
# quandl is deprecated — replaced by nasdaq-data-link
nasdaq-data-link>=1.0.4

# ── NLP & sentiment ────────────────────────────────────────────────────────────
textblob>=0.18.0
vaderSentiment==3.3.2
nltk>=3.9
googletrans==4.0.0rc1

# ── Portfolio & backtesting ────────────────────────────────────────────────────
backtrader==1.9.78.123
# pyfolio is unmaintained; pyfolio-reloaded is the actively maintained fork
pyfolio-reloaded>=0.9.5

# ── Utilities ──────────────────────────────────────────────────────────────────
streamlit>=1.35.0
ipython>=8.24.0
schedule>=1.2.2
websocket_client>=1.6.1
config==0.5.1
yahoo_earnings_calendar==0.6.0
